Feat #21 - 질문 게시글 채택된 답변 있을시 수정,삭제 못하게 변경

+++ src/questions/questions.service.ts
@@ -78,6 +78,11 @@ export class QuestionsService {
     if (result.authorId !== user.id) {
       throw new BadRequestException('작성자만 수정이 가능합니다');
     }
+    if (result.accepteAnswerId) {
+      throw new BadRequestException(
+        '답변이 채택된 게시글은 수정이 불가능합니다.',
+      );
+    }
     await this.questionsRepository.save([
       {
         id: result.id,
@@ -97,6 +102,11 @@ export class QuestionsService {
     /*  question을 user가 만든게 맞는지 check */
     if (result.authorId !== user.id) {
       throw new BadRequestException('작성자만 수정이 가능합니다');
+    }
+    if (result.accepteAnswerId) {
+      throw new BadRequestException(
+        '답변이 채택된 게시글은 삭제가 불가능합니다.',
+      );
     }
     await this.questionsRepository.delete({ id: questionId });
     return true;
